use pages directory for auth
