+++ workspace/apps/fira-app/next.config.js
@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const config = {
   reactStrictMode: true,
-  // pageExtensions: ['page.tsx', 'api.ts'],
+  pageExtensions: ['page.tsx', 'api.ts'],
 };
 
 module.exports = config;
