--- workspace/apps/fira-app/src/AppProviders.tsx
@@ -1,5 +1,0 @@
-import { ChakraProvider, theme } from '@chakra-ui/react';
-
-export function AppProviders({ children }) {
-  return <ChakraProvider theme={theme}>{children}</ChakraProvider>;
-}
