--- workspace/apps/fira-app/src/app/(dashboard)/layout.tsx
@@ -1,9 +1,0 @@
-import { Layout } from 'src/components/layout/Layout';
-
-type DashboardLayoutProps = {
-  children: React.ReactNode;
-};
-
-export default function DashboardLayout({ children }: DashboardLayoutProps) {
-  return <Layout>{children}</Layout>;
-}
