--- workspace/apps/fira-app/src/app/(dashboard)/networth/page.tsx
@@ -1,7 +1,0 @@
-export default function NetWorth() {
-  return (
-    <div>
-      <h1>NetWorth</h1>
-    </div>
-  );
-}
