--- workspace/apps/fira-app/src/app/auth/layout.tsx
@@ -1,9 +1,0 @@
-import { AuthLayout } from './AuthLayout';
-
-type AuthLayoutProps = {
-  children: React.ReactNode;
-};
-
-export default function Layout({ children }: AuthLayoutProps) {
-  return <AuthLayout>{children}</AuthLayout>;
-}
