--- workspace/apps/fira-app/src/app/layout.tsx
@@ -1,20 +1,0 @@
-import { Session } from 'next-auth';
-
-import { Providers } from './providers';
-
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
-};
-
-type RootLayoutProps = { children: React.ReactNode; session: Session };
-
-export default function RootLayout({ children, session }: RootLayoutProps) {
-  return (
-    <html lang="en">
-      <body>
-        <Providers session={session}>{children}</Providers>
-      </body>
-    </html>
-  );
-}
