--- workspace/apps/fira-app/src/app/page.tsx
@@ -1,11 +1,0 @@
-'use client';
-import { Heading } from '@chakra-ui/react';
-
-export default function IndexPage() {
-  return (
-    <div>
-      <Heading>Logged Out</Heading>
-      <pre>{JSON.stringify({}, null, 2)}</pre>
-    </div>
-  );
-}
