+++ workspace/apps/fira-app/src/components/layout/Layout.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import {
   Box,
   Drawer,
