+++ workspace/apps/fira-app/src/pages/api/auth/[...nextauth].api.ts
