+++ workspace/apps/fira-app/src/pages/index.page.tsx
@@ -1,0 +1,7 @@
+export default function Index() {
+  return (
+    <div>
+      <h1>Home</h1>
+    </div>
+  );
+}
