--- workspace/apps/fira-app/src/pages__/settings/account.page.tsx
@@ -1,13 +1,0 @@
-import { Layout } from 'src/components/layout/Layout';
-
-export default function SettingsAccount() {
-  return (
-    <div>
-      <h1>Settings Account</h1>
-    </div>
-  );
-}
-
-SettingsAccount.getLayout = function getLayout(page: React.ReactNode) {
-  return <Layout>{page}</Layout>;
-};
